Separate Consent Manager Builder example
